change doses intengerFiled to decimalField

+++ shop/models.py
@@ -6,7 +6,7 @@ from raw_material.models import WareData, ProductAcquisition
 
 class CoffeeMake(models.Model):
     c_make_ware = models.ForeignKey(ProductAcquisition, on_delete=models.CASCADE, verbose_name='Kávé fajta')
-    c_make_dose = models.IntegerField(verbose_name='Adag', default=4)
+    c_make_dose = models.DecimalField(max_digits=2, decimal_places=1, verbose_name='Adag', default=4)
     c_make_user = models.ForeignKey(
         User, on_delete=models.SET_NULL, 
         null=True, blank=False, 
@@ -18,10 +18,10 @@ class CoffeeMake(models.Model):
 
 class CoffeeOrder(models.Model):
     coffee_selected = models.ForeignKey(CoffeeMake, on_delete=models.CASCADE, verbose_name='Lefoglalt Kávé')
-    coffee_dose = models.IntegerField(verbose_name='Kávé [0,5-2] Adag', default=1)
-    sugar_dose =  models.IntegerField(verbose_name='Cukor [0,5-3] Adag [5 g]', default=1)
-    milk_dose =  models.IntegerField(verbose_name='Tej [0,5-4] Adag [50 ml]', default=1)
-    flavour_dose =  models.IntegerField(verbose_name='Ízesítő [0,5-2] Adag [5 g]', default=1)
+    coffee_dose = models.DecimalField(max_digits=2, decimal_places=1, verbose_name='Kávé [0,5-2] Adag', default=1)
+    sugar_dose =  models.DecimalField(max_digits=2, decimal_places=1, verbose_name='Cukor [0,5-3] Adag [5 g]', default=1)
+    milk_dose =  models.DecimalField(max_digits=2, decimal_places=1, verbose_name='Tej [0,5-4] Adag [50 ml]', default=1)
+    flavour_dose =  models.DecimalField(max_digits=2, decimal_places=1, verbose_name='Ízesítő [0,5-2] Adag [5 g]', default=1)
     coffe_user = models.ForeignKey(
         User, on_delete=models.SET_NULL, 
         null=True, blank=False, 
@@ -29,3 +29,4 @@ class CoffeeOrder(models.Model):
         verbose_name = 'Lefoglalta',
         default=User)
     coffee_reg = models.DateTimeField('Rögzítés dátuma', default=timezone.now)
+
